requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.22.0
sqlalchemy==2.0.22
asyncpg==0.28.0
pydantic==1.10.12
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.2.3
numpy==1.26.4
pytest==7.4.2
httpx==0.24.1
python-multipart==0.0.6
aiofiles==23.1.0

# pip install email-validator
# pip install aiosqlite
# pip install jinja2
# pip install python-multipart

